Add ProfilController and others
