+++ src/main/java/com/carto/sn/controller/ErrorController.java
@@ -1,0 +1,12 @@
+package com.carto.sn.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+public class ErrorController {
+	@RequestMapping("appErreur")
+	public String appErreur() {
+		return "appErreur";
+	}
+}
